Add file option to new document button
